[SYCL][ESIMD] Remove negative value for converted reference data (intel/llvm-test-suite#874)
Remove negative value for case source type is floating-point type and
uint is destination type.
The C++17 draft says the following statement:
A prvalue of a floating-point type can be converted to a prvalue of an
integer type. The conversion truncates; that is, the fractional part is
discarded. The behavior is undefined if the truncated value cannot be
represented in the destination type.