[ESIMD] Fix compilation from static libraries.
sycl-post-link previously removed "llvm.used" global, but that was
not enough as the initializer of "llvm.used" was not deleted and
some elements of initializer used kernels.
As a result the GenXSPIRVWriterAdaptor did not transform the kernel
arguments properly thinking that the kernel was called from inside the module,
which was not true.
Removing the initializer fixed the issue.

Signed-off-by: Vyacheslav N Klochkov <vyacheslav.n.klochkov@intel.com>