[SYCL] Fix get_native for sycl::event (#5765)
Reworked event_impl: added new member flag to observe the state of init of the event and edited getter alongside default event_impl constructor, also made necessary members mutable, so that it supports lazy init, which now occurs when a get_native (both member and free) tries to get implementation of event which is constructed with the default constructor. These changes are to make sycl::event class meet the ad hoc requirements of the specification.