[spirv-to-ir-wrapper] Prevent non-IR from being passed along (#5688)
Introduce an option -skip-unknown-input which can be used to prevent unknown
files from being passed through the tool.  There are usage models involving
archives in which the 'fat archives' do not contain all fat objects.  In
the case where this occurs, provide a way to prevent the non-IR object
from the device side to be passed through as it is not valid at the
device link step.