[Driver][SYCL] Improve -lname static linking with shared objects
By default, the expectation for linking is dynamic.  Abide by these
rules when processing library files for fat static archives.

For example:  clang++ -fsycl -lname

This should only attempt to unbundle libname.a when there is no libname.so
found.  Exception being when we are forcing static linking with options
such as -static and -Wl,-Bstatic.