[SYCL] Implement device_global host-side memory operations
This commit implements the copy and memcpy operations to and from
device_global. If the device_global does not have device_image_scope the
memory operation will be on the underlying USM memory, while if the
operation is on a device_global with device_image_scope the runtime will
try to find a suitable program in the program cache or build a new
program using the image using it.

Signed-off-by: Steffen Larsen <steffen.larsen@intel.com>