[ESIMD] Import and adapt libcxx/include/experimental/simd for invoke_simd. (#5773)
- import libcxx/include/experimental/simd into sycl/include/std/experimental/simd.hpp
and adapt for invoke_simd, so that it can be compiled with SYCL programs. This is
temporary solution until std::simd is supported by the C++ compiler implementations -
MSVC, clang. invoke_simd specification is defined in terms of std::simd and
std::simd_mask classes.
Changes to the original file are marked with the ENABLE_SYCL_EXT_ONEAPI_INVOKE_SIMD
macro.

- Add sanity/compilation test. More will come with invoke_simd implementation.

Signed-off-by: Konstantin S Bobrovsky <konstantin.s.bobrovsky@intel.com>