[SYCL] Pass buffer_location property to buffer (#5604)
There is the accessor property "buffer_location" that allows to allocate buffer in definite location (spec: https://github.com/intel/llvm/blob/sycl/sycl/doc/extensions/supported/sycl_ext_intel_buffer_location.asciidoc)
Current implementation doesn't implement allocating a buffer at the passed location and the buffer will be re-sided when a kernel is enqueued. It leads to problems of various kinds.
The proposed solution implies adding new buffer property in order to store it in a buffer and use it when the buffer is allocated.
If the property is not supported by device it will be ignored.