[Test][SYCL] Add test to verify SPIR kernel argument names (#5819)
This is a follow-up to PR#5772. This test verifies that
user specified names are retained in SPIR kernel argument.
Please note that the names were already retained (prior to
PR#5772) when kernel was specified as a functor. PR#5722 added
code to ensure the same behavior when kernel is specified using
lambda.

Signed-off-by: Elizabeth Andrews <elizabeth.andrews@intel.com>