[Driver] Allow printing defines for sycl when input is stdin
With normal source file as input, clang is able to print defines for sycl.
With stdin as input, however, clang fails to do that due to errors from
the tool append-file, which is because append-file does not allow using
stdin as input.
We need to allow clang to print defines for sycl when input is stdin.

Signed-off-by: Qichao Gu <qichao.gu@intel.com>