[Driver][SYCL] Allow for preprocessing to succeed with coding errors (#5784)
Typical behavior is to allow for preprocessing to succeed regardless of
the input file.  When building with -fsycl, this is not possible due to
the requirement to generate the integration ehader and footer before we do
the actual preprocessing step.

Adjust the compilation when preprocessing is requested to allow for
the device compilation to fail and continue to perform the preprocessing
steps.