[SYCL][CUDA] Don't link pi_cuda against libsycl
The plugins are dynamically loaded and shouldn't be linked against
libsycl.

In this case it meant that a build of libsycl with the CUDA plugin
enabled was unusable on a system where CUDA isn't installed, even when
using other plugins or the host device.

When the plugins are dynamically loaded, if some dependencies are
missing the plugins will just be skipped instead of crashing the
application.