[SYCL][L0] Move command list cache usage under mutex (#5874)
For the example of usage that may cause issue please refer to assert_in_simulteneous_kernels E2E test. n queues that are working on the same device and on default context will work with the same zeCompute/CopyCommandListCache without protection since mutexes for queues are different buwhile cache is the same.
Operator[] may trigger insertion and rehash so simulteneous access is not thread-safe. In this code cache is not prefilled and insertion is done on the first access for every unique ZeDevice what may end up with simulteneous write access and memory corruption

Signed-off-by: Tikhomirova, Kseniya <kseniya.tikhomirova@intel.com>