[SYCL] Store the kernel object size in the integration header (#5862)
Sometimes when the host compiler is different from the device compiler,
it is possible that the size of the kernel object on the host will differ from the
size of the kernel object on the device. One such case is when constexpr
variables are used; compilers differ on whether they need to be captured and
this causes a discrepancy in the size of the kernel object and leads to failures
at runtime.

This change adds an interface to the integration header which returns the size
of the kernel object as computed on the device. The expectation is that the
library will assert the size of a kernel when compiling the host to detect
potential mismatches.