[SYCL][CUDA] Add IPSCCP pass to O0 by default (#5900)
The IPSCCP pass can set branch conditions to ConstInt and swap conditional branches to unconditional branches.
This is necessary at O0 in the nvptx backend in cases where the `nvvm_reflect` function is used: after the nvvm-reflect pass is called, dead branches containing unused instructions aimed at a different architecture generation (SM version) to the one compiled for can remain.

A solution only targeting branches that are using the `nvvm_reflect` function was initially explored by adding a patch to the existing nvvm-reflect pass.  This solution would require considering several cases and was abandoned in favour of a simple comprehensive solution of simply adding the IPSCCP pass to OO: since after discussions it turned out that other backends face a corresponding issue, it was decided that a simple temporary DPC++ solution is favoured and that later on in the year a permanent general solution will be worked on.

New backend flag `use-ipsccp-nvptx-O0` can remove the IPSCCP pass from O0 when set false, at the users discretion.