[SYCL][Doc] Add auto_local_range extension
This extension defines a new auto_range variable which can be used to
define a sycl::nd_range with an unspecified work-group size. If such a
sycl::nd_range object is used to launch a SYCL kernel, an implementation
is free to launch the kernel with any valid work-group size (as defined
by the SYCL specification).

Signed-off-by: John Pennycook <john.pennycook@intel.com>