Make sycl/test/tools/abi_check_positive work with lld linker (#5994)
$ clang --driver-mode=g++ -D_GLIBCXX_ASSERTIONS=1 abi_check_positive.cpp ; nm a.out | grep _dl_relocate_static_pie
0000000000401050 T _dl_relocate_static_pie

$ clang -fuse-ld=ld.lld  --driver-mode=g++ -D_GLIBCXX_ASSERTIONS=1 abi_check_positive.cpp ; nm a.out | grep _dl_relocate_static_pie
00000000002015e0 t _dl_relocate_static_pie

Note, that in order to reproduce it, one has to enable lld as the default linker
on the system and just LLVM_ENABLE_LLD=ON/LLVM_USE_LINKER=LLD isn't enough as
that doesn't change how the linker is invoked inside the test.

Also, relatively recent version of LLD is needed to pass the ABI checks in the
built libsycl.so, e.g. LLD-10 doesn't mark aliases produced via linker
script (abi_replacements_linux.txt) with Function type. LLD-12 doesn't have that
issue.