[SYCL] Make lazily initialized events use default context
Recent changes to get_native made events adhere to default-constructed
events using the default context. This was done through using a
default-constructed context in static storage. However, the statically
stored context could cause the context to live longer than expected.
These changes moves the lazy initialization of context on events with
no context to the get-context method, which is in turn called by the
get-plugin method. With this, events with lazily-initialized native
events, such as default-constructed events, will be associated with the
default context of the platform of the default device.

Co-authored-by: Kharchikov, Igor <igor.kharchikov@intel.com>
Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>