[SYCL] Add specialization of reduction_impl
The specialization allows the type of the reduction (span<T, Extent>)
to be separate from the type of the span (T). T is used to determine
the reduction algorithm that should be used, and to create temporary
storage.

A new static member function "num_elements" is added to all reduction_impl
specializations to distinguish between scalar reductions and array reductions.
A scalar reduction (all existing implementations of reduction_impl) always
has (num_elements == 1); an array reduction (this new specialization) has
(num_elements == Extent), where Extent is the extent of the span.

Signed-off-by: John Pennycook <john.pennycook@intel.com>