[SYCL][NFC] Refactor to avoid code duplication
This commit refactors the reduction implementation to avoid code duplication,
and additionally makes a number of changes to the class hierarchy for future
extensibility.

All functionality and member variables expected to be common to all reductions
is moved to a common base class (reduction_impl_common). The existing
reduction_impl_base is unsuitable for this purpose because it was deliberately
designed not to be a template class.

The reduction_impl is now templated on a reduction algorithm, with any
functionality related to the current (default) algorithm encapsulated in the
default_reduction_algorithm class. This template is carried from the reduction
to any reducers it creates, enabling future specialization of both reduction
and reducer for interesting combinations of type, extent and properties.

The reducer class is simplified using CRTP to avoid duplicate definitions of
combine() and atomic_combine() for scalar and array reductions.

The notion of a reducer's "dimensionality" is now tied to the dimensionality of
the reduction being performed (i.e. 0 for scalars, 1 for spans) and not to the
dimensionality of the input accessor/buffer. This will simplify extending
reductions to true multi-dimensional array reductions (i.e. with md_span).

Signed-off-by: John Pennycook <john.pennycook@intel.com>