Remove more getPointerElementType queries.
Most of this tranche of calls were being used to disambiguate between different
built-ins that share the same or similar names.

The exception is the change in visitCallSPIRVPipeBuiltin, which is a convoluted
way of adding a cast to i8 addrspace(4)* if not already i8 addrspace(4)*, so it
is instead rewritten to check if a cast needs to be added without querying
getPointerElementType.

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/f727f4d