[SYCL][libclc][CUDA] Add --ffast-math support (#5801)
This patch allows the `--ffast-math` compiler
flag to substitute the regular `genfloatf` math
built-ins with their `::native` versions.

Moreover, this patch completes the support of natives
built-ins for `libclc/ptx-nvidiacl` connecting
them with the `__nv_fast` functions present in
libdevice. If a fast function is not available in
libdevice the corresponding `nvvm` intrinsic is used.