[SYCL] Fix enabling of persistent device code cache in unit-tests (#6062)
The persistent device code cache currently only parse the configurations
indicating if it is enabled once, caching the result. This can cause
unit-tests for the cache to fail as they may be run after other tests
that may have populated the cached result. To fix this, the persistent
code cache now exposes a new function forcing the configuration to be
reparsed.
Additionally, the unit-tests will reset the SYCL_CACHE_PERSISTENT after
each test case if it was set during.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>