[SYCL] Make local accessor argument allocation 1-byte minimum (#6032)
Some DPC++ backends, such as OpenCL, does not allow setting local memory
kernel arguments with a size of zero bytes. As such, the runtime library
can fail with a backend error if a local accessor has a zero-size. These
changes make a special case for local accessors with a zero byte size
making it allocate a single byte of memory.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>