[SYCL][CI] Update GPU driver dependencies (#6082)
Note, there is some difference in which docker image is used by the CI that can result in the CPU runtime being at /opt/runtimes or /runtimes. So I'm also expanding the test suite environment to support both paths. That will allow this PR to pass tests, as well as others that may depend on newly made docker images, until we can better understand the reasons for the differences.

Signed-off-by: Chris Perkins <chris.perkins@intel.com>