[Driver][SYCL] Update device compilation with -save-temps
When using -save-temps, we were adding -disable-llvm-passes during the
device LLVM-IR compilation.  This causes the behavior of using -save-temps
to not match without in regards to device compilations.  Make sure that
when we use -save-temps and are performing a SYCL device compilation, we
allow for optimization.

Also do a small refactor of -save-temps tests to consolidate.