Make the LowerBitCast pass support opaque pointers.
This is a relatively large change, as the original pass relied on being able to
track the initial bitcast <3 x i64>* to <6 x i32>* to know where to start
rewriting. Instead, this patch starts at the final invalid extractelement call
and works its way backwards as far as necessary to generate correct code.

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/83b2dfd