[SYCL][L0] Adds device member to L0 make_queue input type
This commit adds a new device member to the make_queue input type for
the L0 backend. This allows the L0 backend to correctly associate the
right device with the command queue to create the PI queue from,
preventing unintented and illegal behavior when the PI queue is later
used for otherwise valid operations.

It is currently still valid to create a SYCL queue from a native L0
command queue without associating the correct device, which will lead
to an unspecified device being selected, similar to the old behavior.
This behavior is however marked as deprecated.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>