[SYCL] Restrict collection of information for optimization record
Collect information for optimization record only if the optmization
record is saved by user (i.e. -fsave-optimization-record or
-opt-record-file is passed). This patch prevents unecessary calls
to opt-report handlers when not required.

I couldn't think of a way to add a test for this since there is no
change in output. Information is collected when the flag is passed
and emitted into optimization record. Information is no longer
collected when there is no optimization record.

This is a follow up to https://github.com/intel/llvm/pull/3492
and https://github.com/intel/llvm/pull/3730/

Signed-off-by: Elizabeth Andrews <elizabeth.andrews@intel.com>