[SYCL] Remove LockedQueue usage which can lead to deadlock
Currently LockedQueue parameter is used several functions for cleaning
resources. It specifies which queue is locked in the caller of these
functions to not lock it twice. Problem is that some of the functions
internally process events in a loop which can be from different queues
and these queues are locked as part of this processing. It leads to the
problem that we can have nested locks of differnt queues and this is a
classic deadlock situation.