[SYCL] Make cleanup() and resetCommandLists() free functions
The resetCommandLists() function performs release of events in command lists
and we release the queue if event's refcount reaches 0, so it is possible
that queue can be destructed inside this function. That's why it should be
free function and not a method of the queue.

Same logic for the cleanup(), it calls EventRelease and event can be
deleted, so it must be free function.