[SYCL] Change online compiler test to build IL based on queue (intel/llvm-test-suite#1028)
Level-zero expects a single device when creating a module. This means
that the online compiler test for level-zero has an assertion on the
supplied context assuring only a single device. However, on multi-GPU
systems this assertion will fail, in turn causing the test to report a
failure. Rather than skipping the test on multi-GPU systems, these
changes makes the test use the device in the queue rather than the first
device in the associated context.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>