[SYCL][CUDA] Add env variable to specify max local mem size (#6173)
Previously the max local mem allocation in DPC++ for CUDA backend was 48KB for most devices.

However as https://docs.nvidia.com/cuda/ampere-tuning-guide/index.html notes, for the A100 the max local mem dynamic allocation is in fact 164KB.

This PR introduces an environment variable SYCL_PI_CUDA_MAX_LOCAL_MEM_SZ which allows you to manually specify the max local memory in bytes allowed to be allocated per kernel for a given application. If an invalid value is specified (one that exceeds the device's capabilities/is negative) then a runtime error will be thrown.

Using:
SYCL_PI_CUDA_MAX_LOCAL_MEM_SZ=166912 ./a.out

Allows the application to use up to 163KB of local memory, if the device supports it.