[SYCL][CUDA][PI] Improve performance of piQueueFinish (#6201)
Improves performance of `piQueueFinish` and therefore of `queue::wait()` on CUDA backend by reducing the number of `cuStreamSynchronize()` calls invoked. This in most use cases fixes the slowdown to  `queue::wait()` introduced in https://github.com/intel/llvm/pull/6102.

This does not change any interface so there are no changes to the test suite.