[CI] Fix Update GPU Driver GitHub Action (#6234)
We seem to be tripping over https://github.com/cli/cli/issues/2979   in the sycl_update_gpu_driver.yml  action. Fortunately, adding the `--head` flag to the PR creation command is trivial and _should_ work around it.  Also, the `sycl` branch is protected  on intel/llvm, so am switching from a direct push to a PR for the nightly.

Note that this has been tested on my fork and works, but that was true for its previous incarnation. It is possible that this fix may yet be incomplete. Won't know until we exercise it.