[ESIMD] Implement stateless memory accesses enforcement
The driver option -fsycl-esimd-stateless-access makes the compiler
to use stateless memory accesses for those ESIMD kernels and ESIMD
intrinsics that use accessors.

The option (if used) defines the macro ESIMD_FORCE_STATELESS_MEM_ACCESS
to map the calls of API using accessors to calls of API using USM pointers.
It also passes a switch to sycl-post-link to signal it that it should
ignore the buffer_t attribute and use svmptr_t.

Signed-off-by: Vyacheslav N Klochkov <vyacheslav.n.klochkov@intel.com>