[SYCL][FPGA] Add support for new attribute for loop pipelining
This patch adds support for new FPGA attribute called [[intel::fpga_pipeline(N)]] that takes a Boolean parameter.
N can be a template parameter or constexpr. If the user didn't provide a parameter, the default value of N would
be 1, which implies enable pipelining.

This attribute can be applied to loops.

Example syntax:

// will disable loop pipelining
[[intel::fpga_pipeline(0)]]
for (int i = 0; i<N; ++i) {...}

LLVM IR:
When applied on a loop, the LLVM IR should be attached to llvm.loop metadata:

!{!"llvm.loop.intel.pipelining.enable", i32 N}

The value of the !"llvm.loop.intel.pipelining.enable" metadata is an i32 value that corresponds to the boolean parameter N.
A value of 0 is used for not pipelining the loop.
A value of 1 is used for pipelining the loop.

If the parameter is not given, value of N will be the default value 1,  which implies we will pipeline the loop, and the IR will look like

!{!"llvm.loop.intel.pipelining.enable", i32 1}

Error Message:

We should emit an error when [[intel::initiation_interval(k)]] and [[intel::fpga_pipeline(0)]] are used together:

"error: initiation_interval and fpga_pipeline(0) attributes are not compatible"

We should emit an error when [[intel::max_concurrency(k)]] and [[intel::fpga_pipeline(0)]] are used together:

"error: max_concurrency and fpga_pipeline(0) attributes are not compatible"

Note: this basically performs very similar functionalities as [[intel::disable_loop_pipelining]], the plan is to deprecate disable_loop_pipelining in future releases and use [[intel::fpga_pipeline(n)]] instead.

Signed-off-by: Soumi Manna <soumi.manna@intel.com>