Address review comments
* Remove limit for commandlists in the cache.
* Protect access to ZeCache with its own mutex.
Access to ZeCache using -> can trigger computation of the property and
we don't know when this is going to happen, so need to use unique_lock
for each access because each access can potentially do write.
* Remove unnecessary device locks considering ZeCache properties are guarded
with their own lock, many data members are const and other members like
SubDevices and QueueGroup vectors are calculated only once during
caching of devices and its subdevices.
* Lock context for shared access in piextUSMDeviceAlloc and
piextUSMHostAlloc.