[SYCL] Fix PersistentDeviceCodeCache unittests
https://github.com/intel/llvm/pull/6256 changed the binary format used
by the PersistentDeviceCodeCache tests to be a parameter, however the
tests would be setting the format after the image had been initialized,
which causes the image to think it is still the old format after
changing the binary format of the PI image to the format read from the
GTest parameter. With these changes the format will be set immediately
from the parameter rather than deferring it to during setup.
Additionally it fixes the SetUp and TearDown overrides.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>