[Driver][SYCL] Update diagnostic behavior for -fsanitize with -fsycl
When using -fsanitize for spir64 targets, the driver would error out not
allowing for compilation to move forward.  -fsanitize is currently not
supported for spir64 targets, but it is supported for host.

Restrict the -fsanitize option to only be used for the Host compilation but
also emit a diagnostic that -fsanitize is being ignored for the target
compilation.