[SYCL] Add implicit rpath to libsycl directory
This patch adds the flag `-f{no-}sycl-implicit-rpath`, which ensures
that on Linux-like platforms the directory containing `libsycl.so` in
the clang build or installation is added to the SYCL binaries' `RPATH`.
This flag is enabled by default.

This makes it easier for SYCL binaries to find `libsycl.so` and removes
the need for user to set `LD_LIBRARY_PATH` in most cases.

This flag is similar to the OpenMP `-fopenmp-implicit-rpath` flag.