[SYCL] Keep used function declarations during llvm.used removal (#6308)
It is possible for a function declaration to be used by both @llvm.used
and, e.g., a spir_func definition. Ensure that we don't attempt to erase
such declarations when removing @llvm.used.

Add a regression test. Most builds will hit an assertion in the first RUN line when the
declaration is incorrectly erased.