[SYCL] Ensure correct errc thrown for unassociated placeholder accessor (4.7.6.9) (#5737)
A simple fix for ensuring correct errc is thrown for an unassociated placeholder accessor (4.7.6.9)

Note that this fix is not complete. It only addresses the most common cases. There are still cases where an unassociated placeholder accessor could get by undetected and crash in the kernel. Unfortunately, the fix for those cases is not trivial. I have the outlines of a fix, but it would be disruptive, too disruptive for such a low priority fix. I will open a ticket about it and discuss it with various stakeholders.

Expanded testing to cover this here: intel/llvm-test-suite#890

Signed-off-by: Chris Perkins chris.perkins@intel.com