[SYCL] Fail on kernel lambda size mismatch
Some host-compilers may create lambdas with captures that do not match
the ones used when extracting kernel descriptors. The compiler does not
currently have a stable way of handling these cases, so instead this PR
adds a static assertion that the sizes match with a message informing
about the limitation.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>