[SYCL][FPGA] Allow safelen = 0 with ivdep attribute (#6336)
This patch allows safelen = 0 with ivdep attribute and emits suppressible warning when safelen of 0 or 1 is used

Current behaviour:

ivdep with safelen of 1 is allowed with no warning. Safelen of 0 causes an error.

Desired behaviour:

safelen of 1 and 0 should both be allowed but will have no effect on the loop.

Both should emit a warning. The warning should be suppressible with a -Wno flag.

Justification:

The safelen parameter of an ivdep will sometimes be set based on a template argument.
It makes this type of code much cleaner if the setting of 0 is permitted.

The warning for both the 0 and 1 case is helpful for users who might misunderstand the
meaning of the safelen parameter and set a value of 1 (or possibly 0) thinking it will have some effect.
The warning should be suppressible for those developers using the templated coding pattern described
above who understand and accept that a safelen of 0 will have no effect.

Signed-off-by: Soumi Manna soumi.manna@intel.com