[SYCL][FPGA] Remove deprecation message for intel::disable_loop_pipelining attribute
https://github.com/intel/llvm/pull/6254 added the fpga_pipeline attribute and
deprecated the intel::disable_loop_pipelining attribute.
While this was in the pipeline, the internal decision was made to switch to properties
rather than attributes for this type of thing.

At this point, internal request is that (at a minimum) the deprecation message for
intel::disable_loop_pipelining be removed, as we no longer plan to deprecate it.

Signed-off-by: Soumi Manna <soumi.manna@intel.com>