[Driver][SYCL] Add support for large device code via a linker script
Add -fsycl-huge-device-code as a driver option. When enabled, the driver
generates a linker script which directs the linker to place device code
later in the binary, which makes it less likely to create a distance
larger than that which a PC32 relocation can span.

For example, if a 3GB __clang_offload_bundle__ section is placed betwen
.text and .rodata, a PC32 relocation to .rodata in the .text is not
possible. With this option enabled the device code will be placed after
both rather than between them.