Fix out of bounds access in SPIRVToOCLBase::mutateArgsForImageOperands
When called from visitCallSPIRVImageReadBuiltIn, Args only has three
elements and ImOpArgIndex is 2. This has been causing this function to
"update" an argument past the end of the vector, which caused a crash
when running transcoding/image_signedness.ll on Windows because the
debug build in Windows has bounds checking on std::vector.

With this change, it should update the correct (and not past the end)
argument instead and satisfy the Windows bounds checks.

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/6904b38