[SYCL] match existing device_impl from context, rather than creating a new one.  (#6423)
The queue_impl(PI) constructor is creating a new device_impl, but that is not strictly correct. Presently this SYCL implementation assumes there is a 1:1 relationship between devicesPI and devic_impl. By creating a new one, there will now be two device_impls (one already in the context) for the same devicePI. In this PR we are simply finding the matching device_impl in the Context and using that rather than creating a new one.

Signed-off-by: Chris Perkins <chris.perkins@intel.com>